enterprise: do not autoscale distributed app
In distributed mode, "app" will be deployed as a StatefulSet in a
consistent hash ring, which is not friendly to autoscaler.

For the time being, we recommend manually scale "app" under distributed
mode instead of using autoscaler.

Note that BuildBuddy Executor is still compatible with autoscaler, this
change is only applicable to "app" deployment.

Closes #23.