executor: make podman the default executor option
This PR flip the default runner in our Executor helm chart from Docker
to Podman. Customers who are using the helm chart usually setup their
deployment in cluster which don't have Docker socket available.
Insted, it's usually a containerd shim or nothing at all.

Podman is the perfect solution for this because it's daemonless.
With Podman, Action containers are run inside the parent Executor
container in a nested fashion. This requires priviledged access to the
parent Executor container/pod, but it's less secure than providing
access to the Docker daemon of the underlying node to Executor.

We have tested Podman runner in our production environment for the past
year and it remains the fastest isolation solution. We are now
recommending this as the default runner for users who self-host their
Executors.